requirements.txt
# Python 3.9+ recommended (see docs/TESTING.md).
boto3>=1.26.0
requests>=2.28.0
requests-aws4auth>=1.1.0
python-dotenv>=1.0.0
